[Feat] CompleteViewController 생성 (#38)
