+++ Uber-iOS/Uber-iOS/Presentation/Reservation/ReservationInfoViewController.swift
@@ -125,7 +125,7 @@ final class ReservationInfoViewController: BaseViewController {
         $0.backgroundColor = .white
     }
     
-    private let goTovehicleReservButton = UIButton().then {
+    private lazy var goTovehicleReservButton = UIButton().then {
         $0.setTitle("차량 서비스 예약", for: .normal)
         $0.applyUberStyle()
         $0.addTarget(self, action: #selector(goTovehicleReservButtonTapped), for: .touchUpInside)
@@ -246,11 +246,8 @@ extension ReservationInfoViewController {
     }
     
     @objc private func goTovehicleReservButtonTapped() {
-        let reservationCompleteVC = UIViewController()
-        let completeView = ReservationCompleteView()
-        completeView.backgroundColor = .white
-        reservationCompleteVC.view = completeView
-        navigationController?.pushViewController(reservationCompleteVC, animated: true)
+        let completeVC = ReservationCompleteViewController()
+        navigationController?.pushViewController(completeVC, animated: true)
     }
 }
